requirements.txt
Flask==2.3.2
Flask-CORS==5.0.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
gTTS==2.5.4
pytesseract==0.3.13
Pillow==11.0.0